requirements.txt
# SheldonOS — Python Dependencies

# ─── Core Framework ───────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
httpx>=0.27.0
asyncio-throttle>=1.0.2

# ─── LLM & AI ─────────────────────────────────────────────────────────────────
anthropic>=0.39.0
openai>=1.54.0

# ─── Databases ────────────────────────────────────────────────────────────────
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.36
alembic>=1.14.0
redis[hiredis]>=5.2.0
neo4j>=5.26.0

# ─── Cognitive Layer ──────────────────────────────────────────────────────────
cognee>=0.1.17
# openviking — install from source: pip install git+https://github.com/volcengine/OpenViking.git

# ─── Workflow & Orchestration ─────────────────────────────────────────────────
# deer-flow — install from source: pip install git+https://github.com/bytedance/deer-flow.git
# ruflo — install from source: pip install git+https://github.com/ruvnet/ruflo.git

# ─── Data & Analysis ──────────────────────────────────────────────────────────
pandas>=2.2.0
numpy>=2.1.0
scipy>=1.14.0
scikit-learn>=1.5.0

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1
pyyaml>=6.0.2
structlog>=24.4.0
rich>=13.9.0
tenacity>=9.0.0
schedule>=1.2.2

# ─── Security ─────────────────────────────────────────────────────────────────
cryptography>=43.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
httpx>=0.27.0  # for TestClient
